+++ lib/karafka/web/ui/controllers/responses/deny.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Web
+    module Ui
+      module Controllers
+        module Responses
+          class Deny
+          end
+        end
+      end
+    end
+  end
+end