+++ lib/karafka/web/ui/pro/app.rb
@@ -134,18 +134,12 @@ module Karafka
                 controller.republish(topic_id, partition_id, offset)
               end
 
-              # Only allow download when enabled
-              if ::Karafka::Web.config.ui.visibility.raw_payload_download
-                r.get String, Integer, Integer, 'download' do |topic_id, partition_id, offset|
-                  controller.download(topic_id, partition_id, offset)
-                end
+              r.get String, Integer, Integer, 'download' do |topic_id, partition_id, offset|
+                controller.download(topic_id, partition_id, offset)
               end
 
-              # Only allow export when enabled
-              if ::Karafka::Web.config.ui.visibility.payload_export
-                r.get String, Integer, Integer, 'export' do |topic_id, partition_id, offset|
-                  controller.export(topic_id, partition_id, offset)
-                end
+              r.get String, Integer, Integer, 'export' do |topic_id, partition_id, offset|
+                controller.export(topic_id, partition_id, offset)
               end
             end