+++ spec/lib/karafka/web/ui/pro/controllers/messages_spec.rb
@@ -60,5 +60,70 @@ RSpec.describe_current do
         expect(response.body).to eq(payload)
       end
     end
+
+    context 'when message exists but downloads are off' do
+      let(:payload) { rand.to_s }
+
+      before do
+        allow(::Karafka::Web.config.ui.visibility.filter)
+          .to receive(:download?)
+          .and_return(false)
+
+        produce(topic, payload)
+        get "messages/#{topic}/0/0/download"
+      end
+
+      it do
+        expect(response).not_to be_ok
+        expect(response.status).to eq(403)
+      end
+    end
+  end
+
+  describe '#export' do
+    context 'when we want to export message from a non-existing topic' do
+      before { get 'messages/non-existing/0/1/export' }
+
+      it do
+        expect(response).not_to be_ok
+        expect(response.status).to eq(404)
+      end
+    end
+
+    context 'when message exists' do
+      let(:payload) { rand.to_s }
+      let(:expected_file_name) { "#{topic}_0_0_payload.json" }
+      let(:expected_disposition) { "attachment; filename=\"#{expected_file_name}\"" }
+
+      before do
+        produce(topic, payload)
+        get "messages/#{topic}/0/0/export"
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(response.headers['content-disposition']).to eq(expected_disposition)
+        expect(response.headers['content-type']).to eq('application/octet-stream')
+        expect(response.body).to eq(payload)
+      end
+    end
+
+    context 'when message exists but exports are off' do
+      let(:payload) { rand.to_s }
+
+      before do
+        allow(::Karafka::Web.config.ui.visibility.filter)
+          .to receive(:export?)
+          .and_return(false)
+
+        produce(topic, payload)
+        get "messages/#{topic}/0/0/export"
+      end
+
+      it do
+        expect(response).not_to be_ok
+        expect(response.status).to eq(403)
+      end
+    end
   end
 end